+++ helper/tx.go
@@ -69,7 +69,7 @@ func GenerateAuthObj(client *ethclient.Client, address common.Address, data []by
 	auth = bind.NewKeyedTransactor(ecdsaPrivateKey)
 	auth.GasPrice = gasprice
 	auth.Nonce = big.NewInt(int64(nonce))
-	auth.GasLimit = uint64(gasLimit) // uint64(gasLimit)
+	auth.GasLimit = gasLimit
 
 	return
 }
